Validações separada e altura usuario corrigido.

+++ tcc.DTO/Usuario.cs
@@ -14,7 +14,7 @@ namespace tcc.DTO
         public DateTime nascimento { get; set; }
         public string sexo { get; set; }
         public Decimal peso { get; set; }
-        public int altura { get; set; }
+        public Decimal altura { get; set; }
         public string objetivo { get; set; }
         public string id_gym_personal { get; set; }
         public string id_nutricionista { get; set; }
